requirements.txt
Django==3.2.8
gunicorn==20.1.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.5